Job is added to release a new DOMjudge version
The release script is stored in domjudge-scripts. For instructions on
which steps to take one should check the howto in:

- https://github.com/DOMjudge/domjudge-scripts/blob/master/new_release_howto.txt

The build will be signed after building and tested and if an
installation works will be uploaded to the DOMjudge.org server.

The GPG key can be found in the gitlab settings in the variables
setting. The SSH key for the rsync is in the same location.

The job is added in another file to keep it readable. Jobs can be
collapsed.